SuccessPCpyqt_visualization/requirements.txt
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyserial>=3.5
joblib>=1.1.0
matplotlib>=3.5.0